Remove the SBS_CANTSENDMORE checks from sbtls_boring_fixup().
The old code seemed to assume that if this flag is set, the packets were
going to be dropped, so there was no need to fix them.  However,
SBS_CANTSENDMORE just prevents userland from enqueueing more data to the
socket buffer.  These packets are already in the socket buffer as NOTREADY
data, so they are going to be transmitted to the protocol and onto the wire
and should thus always be fixed up.