First version of new structures for sessions and ops.
These flatten the structures to assume at most 1 cipher and 1 auth and
remove the length lists.  The order is now specified by a "mode" in the
session.  The cryptop no longer includes duplicate information like alg
and klen values.  Instead, drivers should cache the values from the
session params in the driver-specific session data if needed.