Get crypto.c to compile.
This was actually fairly simple as most of the messiness with the
crd/cri lists is in drivers and consumers.

There is a bit of an XXX about the whole session restore thing.  It is
probably now better as we save a copy of the csp, but we probably need
to think about the key pointers.  I think crypto_newsession should
perhaps save copies and we should make rekeying explicit?