Add a custom implementation of cpu_lock_delay() for x86.
Avoid using DELAY() since it can try to use spin locks on CPUs without a
P-state invariant TSC.  For cpu_lock_delay(), always use the TSC if it
exists (even if it is not P-state invariant) to delay for a microsecond.
If the TSC does not exist, read from I/O port 0x84 to delay instead.

PR:		228768