Checkpoint work to make TLS session state its own separate entity.
This is still a WIP with many things still needing updates.  Untangling
the so refcount again is also going to require reverting some changes I
made previously.  I am also kind of replacing SB_TLS_IFNET with checks
to see if the sb_tls_crypto function pointer is NULL or not instead.