Checkpoint work to cleanup socket reference handling.
This goes back to the pre-IFNET_TLS state where callers of sbtls_enqueue()
donate a single reference on the socket that is released by stbls_encrypt()
at the end.

This also fixes the ext_pgs free handling to drop the sbtls_session
reference.  It does "cheap" drops that don't drop from 1 to 0 inline
and only defers potential final drops to the worker thread.  We could
perhaps do the 1 to 0 inline as well, but m_free() gets called in all
sorts of places and cipher teardown might have crazy work to do.

Also started on various compile fixes.