Populate the full and final TLS header before passing mbufs to crypto.
Specifically, set the real length in the TLS header in sbtls_frame()
along with the nonce for CBC when using explicit IVs.  For GCM we don't
know the sequence number until later in sbtls_seq, so copy the sequence
number into the hader in sbtls_seq().  This should ensure that the
TLS header is fully populated once the mbuf is visible to crypto backends.
This also fixes an issue with retransmits for CBC with hardware offload
where the same nonce (IV) needs to be used during retransmits.