Cache the state of TH_FIN to avoid a use after free.
If a TLS record doesn't have TH_FIN set, then the last sbtls_tls_write_wr
will schedule the mbuf chain to be free'd when the last TLS record
descriptor is harvested by a tx completion.  However, sbtls_write_wr was
then checking TH_FIN in the TCP header after this function returned to
determine if a FIN-only packet should be sent.  With preemption, there is
a race where the mbuf might be reused and that the TCP header in the new
mbuf might have TH_FIN set, but we would try to send it using the wrong
sequence number.  Fix this by reading TH_FIN once before scheduling the
mbuf to be freed.  Note that in the case that TH_FIN being set, we don't
don't schedule the mbuf to be freed until the FIN-only packet descriptor
is completed.