Rework locking of TLS session fields in the socket buffer.
The socket option is really a socket buffer level operation and doesn't
require the INP lock to be held.  Instead, we need to ensure that
setting (or in the future, changing) a TLS session is atomic with respect
to attempts to write more data into a socket buffer.  To that end, use
sblock() to protect the sb_tls_info pointer.
- Don't hold the INP lock around sbtls_crypt_tls_enable.  As a result,
  switch to using M_WAITOK and regular copyin() calls.
- Use sblock() and sbunlock() at the end of sbtls_crypt_tls_enable
  when setting the session in the socket buffer.
- Read the TLS session and deal with TLS_SET_RECORD_TYPE control
  messages after sblock() in sosend_generic.
- In the T6 backend, lock the INP before accessing fields in tp and
  hold it while accessing fields in the tp.  To limit how long it is
  held using a bool to cache the 'using_timestamps' field in cipher.
  Use M_WAITOK in a place in the T6 backend now as well.