Eliminate tls_control.
Instead, set atomic to one for non-app-data TLS records.  This fixes a
potential deadlock if a non-app-data TLS record was sent that was
larger than the socket buffer's hi water mark.  The send would never
have gotten past the one place that was checking tls_control, but
because the other check was missing tls_control, it skipped the
EMSGSIZE error and just slept forever.