Update the fragment cases I missed earlier in ip*_output.
To reduce code duplication, add inline functions in each file to
handle applying a rate limit tag.  The inline functions in the two
files are identical and thus still duplicated, but now there are 2
copies rather than 4.

In the case of fragments, move the check earlier to work with the
existing loop flow.  This has the advantage of not bumping the
counters for packets we aren't going to pass on to if_output.