Fix lagg to allocate a new tag on an ifp mismatch.
Right now this keeps the active tags in a linked list.  This keeps the
tags for the old port around until the socket goes away to avoid issues
with reference counts (a reference to the lagg tag "holds" the child
tags).  However, new tags are always placed at the head of the list
optimizing for the failover case.

The tag mutex is used to prevent races with a modify changing the
pacing rate or attempts to add the same child tag concurrently.