Add a refcount on send tags.
This closes use after free races if there are pending packets in driver
tx rings after the socket is closed (e.g. from tcpdrop).  The cost is
an atomic per sent packet to bump the refcount and an atomic in
m_free to drop it.

The wrapper tags for lagg and vlan "hand-off" the refcount from the
wrapper tag to the child tag.

In order for m_free to work reliably, add a new CSUM_SND_TAG flag in
csum_flags to indicate 'snd_tag' is set (rather than 'rcvif').