Fix some issues with cxgbe and send tag reference counting.
cxgbe was relying on snd_tag_free being called when the inp was detached
so that it could kick the firmware to flush any pending work on the
flow.  This is because the driver doesn't require ACK messages from the
firmware for every request, but instead does a kind of manual interrupt
coalescing by only setting a flag to request a completion on a subset
of requests.  If all of the in-flight requests don't have the flag when
the tag is detached from the inp, the flow might never return the credits.
The current snd_tag_free command issues a flush command to force the
credits to return.  However, the credit return is what also frees
the mbufs, and since those mbufs now hold references on the tag, this
meant that snd_tag_free would never be called.

To fix, explicitly drop the mbuf's reference on the snd tag when the mbuf
is queued in the firmware work queue.  This means that once the inp's
reference on the tag goes away and all in-flight mbufs have been queued
to the firmware, tag's refcount will drop to zero and snd_tag_free will
kick in and send the flush request.  Note that we need to avoid doing this
in the middle of ethofld_tx(), so the driver grabs a temporary reference
on the tag around that loop to defer the free to the end of the function
in case it sends the last mbuf to the queue after the inp has dropped its
reference on the tag.

This also fixes a similar recursion issue in the FW4_ACK handler which
was freeing the completed mbufs when credits were returned while holding
the tag lock.  Now the reference is dropped earlier and the m_freem in
the handler doesn't try to invoke snd_tag_free.