Improve MBUF_EXT_PGS_ASSERT.
Include __func__ and mbuf pointer in panic message.

While here, also clarify the panic message as the old one implied that
M_EXT was known to be set and only the ext type was wrong.  This new
wording was adopted from the inlined version of this in mb_free_mext_pgs.