Restructure to match sbcompress().
Try to make the conditions for compression in sbready_compress() match
sbcompress() as much as possible.

Add a note to explain why 'm' and 'n' are reversed.  One possible bug
is that the old code was checking EOR on the packet being merged instead
of the other way around.  There might also be a bug with needing to
propagate M_EOR from 'n' to 'm' when merging.