Tweak the unmapped compression in sbcompress.
Move it next to where we do mbuf coalescing so codewise it is in the
same block as as coalescing instead of the middle of the block that
adds 'm' to the socket buffer.  This I think makes it obvious that we
don't need to adjust sb accounting since 'm' is clearly not in the
socker buffer yet when it is converted.