Rework event reporting.
Remove the TAILQ of pending events and go back to a single
stopped_vcpu.  The handlers for swbreak and mtrap will now loop trying
to report their event.  For breakpoints, discard the event if the
breakpoint is no longer present when the vCPU thread resumes.

To cope with stepping when there is a pending breakpoint, don't enable
stepping until a vCPU thread resumes after reporting its event.