Remove use page_count for ext_pgs queued for SW TLS.
When M_NOMAP was upstreamed, the 'enc_cnt' member was removed under
the assumption that 'nrdy' could just be used instead since sendfile()
is already done with waiting for I/O at that point.  However, this
assumption was not quite correct.  The 'enc_cnt' member held the count
of all of the pages in a chain.  However, this optimization is also
not really needed.  The only time it was ever used was if encryption
failed to avoid computing the number of pages passed to
mb_free_notready.

Instead, simplify the code to remove enc_count from sbtls_frame and
sbtls_enqueue.  For the extremely common case that encryption
succeeds, npages already holds the page count passed to pru_ready.
For the error case, just walk the remaining portion of the mbuf chain
to compute a full npages value to pass to mb_free_notready().