Fix a regression with removing enc_cnt.
sbtls_enqueue needs to be careful to only encrypt the chain of mbufs
that 'top' held when it was inserted in the socket buffer.  However,
by the time sbtls_enqueue runs, those mbufs might be in the socket
buffer and other mbufs might be appended after the original end.
We have to make sure we don't encrypt those future mbufs as they
might already be in the queue to be encrypted.