Simplify interrupt handling for smc(4).
Rather than using a filter and three separate tasks along with toggling
interrupt masks, simply perform all of the normal work in an ithread
handler like other common 10/100 drivers.  smc(4) is a bit unusual in
that it uses PIO, so some extra state is required on TX to allocate
space in the PIO window for the next packet along with knowing when the
window has room for another packet.  As a result, the interrupt handler
can still modify the interrupt mask, but now it only does so for TX
events related to window management.