Convert ccp(4).
In addition to converting, I tried to fix some issues in this driver
where it would fail if multiple requests were sent on a single session.
Right now it enforces one request per session which is not great.
Allocating per-request space for temporary buffers for the IV, and hash
results would permit multiple requests in flight.