A hack for T6 KTLS.
The T6 TLS makes use of the TOE engine to handle TCP segmentation.
However, due to the way it works, it can only apply a relative adjustment
to the base clock used for TCP timestamps in the upper 4 bits.  The
driver refuses to offload connections with an adjustment that has any
of the low 28 bits set in the offset.  This hack clears the low 28 bits
so that T6 can offload all connections.