Fix another bug in the KTLS write_sgl routine.
Don't clobber the last SGL length if there is an odd number of segments
greater than 1.  We only need to set the "extra" length to 0 if the
last length set was even (j is the last index set at this point, not
the next index to set as in some of othe other write_sgl routines).