Axe per-op key lengths.
Key lengths are now set only by the session params and are fixed for
the lifetime of a session.  Per-op keys must have the length of the
session.

As part of this, don't always require an auth key for keyed digests
when the session is created.  Instead, just check that the key length
is valid.  This avoids geli having to pass in a bogus key to setup
auth.