Copy out aux args after the argv and environment vectors.
Partially revert r354741 and r354754 go back to allocating a fixed-size
chunk of stack space for the auxiliary vector.  Keep sv_copyout_auxargs
but change it to accept the address at the end of the environment
vector as an input stack address and no longer allocate room on the
stack.  It is now called at the end of copyout_strings after the
argv and environment vectors have been copied out.

This should fix a regression in r354754 that broke the stack alignment
for newer Linux amd64 binaries (and probably broke Linux arm64 as well).