Further flesh out CSP_MODE_MTE including /dev/crypto first pass.
I had missed finishing updates to /dev/crypto.  Also, in the
cryptosoft bits I had assumed the payload length spanned the digest
and trailer.  Instead, add a crp_padding_length field so the caller
controls the length of the padding explicitly and have the payload
length really be the payload.