Permit variable IV lengths for requests within a crypto session.
Replace the per-session csp_ivlen field with a per-request
crp_iv_length field.  In practice, this only matters for AES-CCM
which supports nonce lengths from 7 to 13.  The IV length in the
existing CIOCCRYPTAEAD request can be used to request different
IV lengths.

AES-GCM also supports varying IV lengths, but the GHASH optimizes
the case of an IV of 12 bytes and most other crypto drivers in the
tree that support AES-GCM only support the 12 byte case.  Simplify
some drivers to assume an IV of 12 bytes only and assert that GCM
requests only use 12 byte IVs in crp_sanity().