Allocate icl separately from the bhs mbuf.
Allow both the PDU and any external mbufs associated with ICL_NOCOPY
buffers hold a reference on the icl.  The PDU's reference is dropped
by icl_cxgbei_conn_pdu_free(), and the external mbufs references are
released when the mbufs are freed.  This allows the icl to stay around
if needed for ICL_NOCOPY buffers and instead be freed when the last
mbuf is freed.

This removes the need for the hack in rqdrop_locked().