Fix support for interrupt moderation in the initiator.
The previous moderation only worked for the target case and assumed
that a sequence only had a single burst such that it could use the
DataSN in the last BHS as the count of segments and compute the
proper buffer offset from that.  However, this doesn't work on the
initiator side where a single transaction might generate multiple
sequences.  Instead, create a new cxgbei_cmp structure to track the
next expected buffer offset for a given transfer.  Store these
structures in a hash table indexed by the initiator or transfer tags.
For CPL_RX_ISCSI_CMP, check the BHS opcode to determine which tag
to use and then lookup the associated CMP data.  Use the expected
buffer offset to determine how much additional DDP data was placed
and how many segments it spanned.  Bursts are still passed up to
the stack as a single, large PDU, but with the DataSNs rewritten
to appear in ascending order.

Note that for a multi-transfer target data out received in response
to an R2T, the buffer offset begins after the first burst rather
than at 0, and that the cxgbei_cmp structure is reused along with
the page pods.

While here, fix a resource leak if a target transfer is aborted
due to the connection being shutdown.