iSCSI: Add support for segmentation offload for hardware offloads.
Similar to TSO, iSCSI segmentation offload permits the upper layers to
submit a "large" virtual PDU which is split up into multiple segments
on the wire.  Each segment is itself a PDU, so the DataSN field must
be updated to reflect the number of segments transmitted.  The target
layer computes this by dividing the total length by the backend's
ic_max_send_data_segment_length.