cxgbei: Queue work requests to write page pods in the ulp_pduq.
Previously the page pods work requests were written to the connections
offload TX queue immediately, now they are placed in the ulp_pduq
behind any other pending work requests.

XXX: I'm not sure why we need this change.  The comments in the
earlier patches implied that the original code was using the ctrlq
instead of of the offload TX queue and these changes fixed that
(presumably to close a race between the R2T going out over the offload
TX queue before the page pods were configured), but the existing code
was already using the offload TX queue.