cxgbei: Support iSCSI offload on T6.
T6 makes several changes relative to T5 for receive of iSCSI data
buffers.

First, earlier adapters issue either 2 or 3 messages to the host for
each PDU received: CPL_ISCSI_HDR contains the BHS of the PDU,
CPL_ISCSI_DATA (when DDP is not used for zero-copy receive) contains
the PDU data as buffers on the freelist, and CPL_RX_ISCSI_DDP with
status of the PDU such as result of CRC checks.  In T6, a new
CPL_RX_ISCSI_CMP combines CPL_ISCSI_HDR and CPL_RX_ISCSI_DDP.  PDUs
which are directly placed via DDP only report a single
CPL_RX_ISCSI_CMP message.  PDUs received on the free lists are
reported as CPL_ISCSI_DATA followed by CPL_RX_ISCSI_CMP.

Supporting this requires changing the CPL_ISCSI_DATA handler to
allocate a PDU structure if it is not preceded by a CPL_ISCSI_HDR as
well as support for the new CPL_RX_ISCSI_CMP.

Second, when using DDP for zero-copy receive, T6 will only issue a
CPL_RX_ISCSI_CMP after a burst of PDUs have been received (indicated
by the F flag in the BHS).  In this case, the CPL_RX_ISCSI_CMP can
reflect the completion of multiple PDUs and the BHS and TCP sequence
number included in the message are from the last PDU received in the
burst.  Notably, the message does not include a count of the number of
PDUs received.  Instead, the driver must track the amount of data
already received for a given transfer and use this to compute the
amount of data received in a burst and the number of PDUs the data
spanned.  In addition, the iSCSI layer currently has no way to permit
receiving a logical PDU which spans multiple PDUs.  Instead, the
driver presents each burst as a single, "large" PDU to the iSCSI
target and initiators.  This is done by rewriting the buffer offset
and data length fields in the BHS of the final PDU as well as
rewriting the DataSN so that the received PDUs appear to be in order.

To track all this, cxgbei maintains a hash table of 'cxgbei_cmp'
structures indexed by transfer tags for each offloaded iSCSI
connection.  When a SCSI_DATA_IN message is received, the ITT from the
received BHS is used to find the necessary state in the hash table,
whereas SCSI_DATA_OUT replies use the TTT as the key.  The structure
tracks the expected starting offset of the next burst as well as the
rewritten DataSN value used for the previously received PDU.