cxgbe tom: Permit rcv_nxt mismatches on FIN for iSCSI connections.
The remote peer might send a FIN in the middle of a burst of data
PDUs.  In the case of T6 with data PDU completion moderation, the
driver would not have seen these PDUs since the final PDU in the burst
was never received resulting in a stale rcv_nxt when the FIN is
received.

Reported by:	Jithesh Arakkan @ Chelsio